added post installation script

+++ ArchHyprland/install.sh
@@ -40,13 +40,8 @@ clear
 
 figlet -f big "DOTFILES"
 
-if gum confirm "Do you want to start the dotfiles installation?" ;then
-    echo -e "\nLet's go.\n"
-elif [ $? -eq 130 ]; then
-        exit 130
-else
-    echo -e "\nUpdate canceled."
-    exit;
+if ! gum confirm "Do you want to start the dotfiles installation?" ;then
+    exit 0
 fi
 
 echo -e "\n\nPress [ENTER] to continue..."
@@ -126,12 +121,11 @@ clear
 
 figlet -f big "CONFIGURATION FILES"
 
+pictures_folder=$(xdg-user-dir PICTURES)
+
 cp -r ./CustomConf/config/* ~/.config
 cp ./CustomConf/.bashrc ~/.bashrc
-cp -r Wallpapers/ ~/Pictures
-
-~/.config/dotfiles/scripts/set_wallpaper.sh ~/Pictures/Wallpapers/wallpaper_1.png &> /dev/null
-~/.config/dotfiles/scripts/set_gtk.sh &> /dev/null
+cp -r Wallpapers/ $pictures_folder
 
 echo -e "\n\nPress [ENTER] to continue..."
 read
@@ -146,24 +140,40 @@ clear
 
 figlet -f big "SDDM"
 
-# Enable sddm
-if [ -f /etc/systemd/system/display-manager.service ]; then
-    sudo rm /etc/systemd/system/display-manager.service
+downloads_folder=$(xdg-user-dir DOWNLOAD)
+
+echo -e "\n\nDo you want to use SDDM?"
+answer=$(gum choose "Yes" "No")
+if [ "$answer" == "Yes" ]; then
+    # Enable sddm
+    if [ -f /etc/systemd/system/display-manager.service ]; then
+        sudo rm /etc/systemd/system/display-manager.service
+    fi
+
+    sudo systemctl enable sddm.service
+
+    if [ ! -d /etc/sddm.conf.d/ ]; then
+        sudo mkdir /etc/sddm.conf.d
+    fi
+
+    sudo cp ~/.config/sddm/sddm.conf /etc/sddm.conf.d/
+
+    echo -e "\n\nDo you want to use sugar-candy as SDDM theme?"
+    answer=$(gum choose "Yes" "No")
+    if [ "$answer" == "Yes" ]; then
+
+        # Download the theme
+        wget -P $downloads_folder/sddm-sugar-candy https://github.com/Kangie/sddm-sugar-candy/archive/refs/heads/master.zip
+        unzip -o -q $downloads_folder/sddm-sugar-candy/master.zip -d $downloads_folder/sddm-sugar-candy
+
+        # Set the theme
+        sudo cp -r $downloads_folder/sddm-sugar-candy/sddm-sugar-candy-master /usr/share/sddm/themes/sugar-candy
+        rm -rf $downloads_folder/sddm-sugar-candy
+
+        # Update the wallpaper
+        ~/.config/dotfiles/scripts/set_sddm_wallpaper.sh $pictures_folder/Wallpapers/wallpaper_1.png &> /dev/null
+    fi
 fi
-
-sudo systemctl enable sddm.service
-
-if [ ! -d /etc/sddm.conf.d/ ]; then
-    sudo mkdir /etc/sddm.conf.d
-fi
-
-sudo cp ~/.config/sddm/sddm.conf /etc/sddm.conf.d/
-
-wget -P ~/Downloads/sddm-sugar-candy https://github.com/Kangie/sddm-sugar-candy/archive/refs/heads/master.zip
-unzip -o -q ~/Downloads/sddm-sugar-candy/master.zip -d ~/Downloads/sddm-sugar-candy
-sudo cp -r ~/Downloads/sddm-sugar-candy/sddm-sugar-candy-master /usr/share/sddm/themes/sugar-candy
-
-~/.config/dotfiles/scripts/set_sddm_wallpaper.sh ~/Pictures/Wallpapers/wallpaper_1.png &> /dev/null
 
 echo -e "\n\nPress [ENTER] to continue..."
 read
@@ -181,8 +191,11 @@ figlet -f big "NVIDIA"
 # Ask if the user is using a proprietary NVIDIA card
 echo -e "\n\nAre you using an NVIDIA GPU?"
 answer=$(gum choose "Yes" "No")
+echo -e "\n\nSetting environment variables"
 if [ "$answer" == "Yes" ]; then
-    echo # TODO
+    cat ~/.config/hypr/configs/environments/nvidia.conf > ~/.config/hypr/configs/environment_vars.conf 
+else
+    cat ~/.config/hypr/configs/environments/default.conf > ~/.config/hypr/configs/environment_vars.conf
 fi
 
 echo -e "\n\nPress [ENTER] to continue..."
@@ -200,45 +213,51 @@ figlet -f big "PROGRAMS"
 
 # Ask if the user wants to download programs
 if gum confirm "Do you want to download some programs (Visual studio, Discord, etc...)? You can select which one." ;then
-    echo
-else
-    clear
-    figlet -f big COMPLETED
-    exit 0
+
+    # Use gum to create a list where the user can pick which packages to install
+    echo -e "\n\nPress [X] on the keyboard to select/unselect and then [ENTER] to confirm.\n"
+    packages=$(gum choose --no-limit "Visual Studio Code" "Bitwarden" "Vesktop (Discord for Wayland)")
+
+    # Output the selected packages
+    echo "Selected packages: $packages"
+
+    # Save original IFS
+    OLD_IFS=$IFS
+
+    # Use IFS to handle multiple selections properly
+    IFS=$'\n'
+
+    # Install the selected packages
+    for package in $packages; do
+        echo "Installing $package..."
+        case "$package" in
+            "Visual Studio Code")
+                install visual-studio-code-bin yay
+                ;;
+            "Bitwarden")
+                install bitwarden yay
+                ;;
+            "Vesktop (Discord for Wayland)")
+                install vesktop yay
+                ;;
+        esac
+    done
+
+    IFS=$OLD_IFS
+
 fi
-
-# Use gum to create a list where the user can pick which packages to install
-packages=$(gum choose --no-limit "Visual Studio Code" "Bitwarden" "Vesktop (Discord for Wayland)")
-
-# Output the selected packages
-echo "Selected packages: $packages"
-
-# Save original IFS
-OLD_IFS=$IFS
-
-# Use IFS to handle multiple selections properly
-IFS=$'\n'
-
-# Install the selected packages
-for package in $packages; do
-    echo "Installing $package..."
-    case "$package" in
-        "Visual Studio Code")
-            install visual-studio-code-bin yay
-            ;;
-        "Bitwarden")
-            install bitwarden yay
-            ;;
-        "Vesktop (Discord for Wayland)")
-            install vesktop yay
-            ;;
-    esac
-done
-
-IFS=$OLD_IFS
 
 echo -e "\n\nPress [ENTER] to continue..."
 read
 clear
 
 figlet -f big COMPLETED
+
+touch ~/.config/dotfiles/cache/do_post_install
+
+echo -e "\n\nDo you want to reboot the system now?"
+if ! gum confirm "" ;then
+    exit 0    
+fi
+
+reboot
